BTDB2 Eco Calculator
This is a Bloons Tower Defense Battles 2 Economy calculator, or BTDB2 for short. This code is not very serious, it's mostly for fun. Each Number will be prefaced with what it represents. The matplotlib projection will show what it would look like if you started with $100, 50eco, and if you constantly spent it for 20 rotations( a rotation is one eco cycle where it gives you money).

More Context:
All of the functions will run once just to show what they do but you will have to modify the file if you want to do something different. Most of it is just spitballing to see what I can do and find when working with the Bloons Tower Defense Battles 2 eco system. I hope one day I can make this more polished and worth people's time but for now this is the minimum viable product. If you know some python and/or have played BTDB2 I would love to get your opinion. Any and all criticism is allowed because I genuinely don't know if my code is just spaghetti or what and I want to grow. Also I know that the raw numbers I give for the bloons might be off, the game has and will continue updating and I care less about the exact cost of each bloon for now and more about whether of not my code works as expected.